Moved Extensions/Install.py fucntions to setuphandlers, kept BBB import for activatePluginInterfaces since this is imported by borg.localrole.
